requirements.txt
flask
pypdf
sentence-transformers
numpy
groq
python-dotenv
